# docs/implementation/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/lessons": {
      "get": {
        "operationId": "PrivateController_lessons",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      },
      "post": {
        "operationId": "PrivateController_createLesson",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Private"
        ]
      }
    },
...
    "/api/v1/editor/articles/{id}/revisions": {
      "post": {
        "operationId": "EditorController_revise",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Editor"
        ]
      }
    },
    "/api/v1/editor/revisions/{id}/submit": {
      "post": {
        "operationId": "EditorController_submit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Editor"
        ]
      }
    },
    "/api/v1/editor/revisions/{id}/reviews": {
      "post": {
        "operationId": "EditorController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Editor"
        ]
      }
    },
    "/api/v1/editor/revisions/{id}/publish": {
      "post": {
        "operationId": "EditorController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Editor"
        ]
      }
    },
    "/api/v1/editor/articles/{id}/withdraw": {
      "post": {
        "operationId": "EditorController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Editor"
        ]
      }
    },
...
    "/api/v1/quizzes/{id}": {
      "get": {
        "operationId": "LearningController_quiz",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/quizzes/{id}/attempts": {
      "post": {
        "operationId": "LearningController_attempt",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
...
    "/api/v1/lessons/{id}/shares": {
      "post": {
        "operationId": "LearningController_share",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
    "/api/v1/shares/{id}": {
      "get": {
        "operationId": "LearningController_shared",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      },
      "delete": {
        "operationId": "LearningController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Learning"
        ]
      }
    },
...
  },
  "info": {
    "title": "HumanScope REST API",
    "description": "Foundation route inventory. Body validation is defined by Zod contracts; response schemas and full OpenAPI coverage remain a release gate.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hs_session"
      }
    },
...
  }
}